# application/services/custom_services/invoice/src/swagger/swagger.yaml
---
openapi: 3.0.0
info:
   version: 1.3.0
   title: appointschedule open api specification
   description: invoice API
   termsOfService: http://api_url/terms/
   contact:
      name: ContactName
      email: hello@wolox.co
      url: https://www.wolox.com.ar/
   license:
      name: Apache 2.0
      url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
   - url: http://localhost:8012
     description: invoice
tags:
- name: invoice
paths:
   "/invoice/{id}":
      delete:
         tags:
         -  invoice
         description: GpDelete
         parameters:
         -  in: "path"
            description: GpDelete
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  invoice
         description: GpGetEntityById
         parameters:
         -  in: "path"
            description: GpGetEntityById
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/invoice/get/search":
      get:
         tags:
         -  invoice
         description: GpSearch
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/invoice/get/update":
      put:
         tags:
         -  invoice
         description: GpSearchForUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/invoice"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/invoice":
      put:
         tags:
         -  invoice
         description: GpUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/invoice"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  invoice
         description: GpGetAllValues
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      post:
         tags:
         -  invoice
         description: GpCreate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/invoice"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/invoice/userid/created_by":
      get:
         tags:
         -  invoice
         description: GpGetNounCreatedBy
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
components:
   schemas:
      invoice:
         type: object
         properties:
            ids:
               type: "string"
               description: Description
            appointments:
               type: "string"
               description: Description
            issued:
               type: "string"
               description: Description
            numbers:
               type: "string"
               description: Description
            statuss:
               type: "string"
               description: Description
            totalamount:
               type: "number"
               description: Description